Fix a destructing issue.
When PHP is destructing the run, some objects are destructed before this
one. And when they are destructured, there is an undetermined
side-effect that remove this static attribute. So we avoid the error by
checking the attribute still exists.

This is not very clean, but fixing this would require to write the
destructors for almost all objects in atoum in order to control the way
objects destruct themselves. I reckon this is a bug in PHP, but very
hard to reproduce in an isolate way. So this is better this way I guess.